pyproject.toml
[project]
name = "orion"
version = "0.1.0"
description = "Orion multi-agent AI orchestration platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"asyncpg>=0.30.0",
	"fastapi>=0.115.0",
	"httpx>=0.28.0",
	"networkx>=3.4.2",
	"pydantic-settings>=2.8.1",
	"redis>=5.2.1",
	"sqlalchemy>=2.0.38",
	"uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
	"ruff>=0.11.0",
]

[tool.uv]
package = false
